Svent 4 to 5 and vite 4 to 6 migration, removed moment.js

+++ vite/clean-dist-cdn.js
@@ -1,13 +1,16 @@
 import fs from 'fs';
 import path from 'path';
+import { fileURLToPath } from 'url';
 
-// Convert the import.meta.url to a URL object to handle file protocol
-const currentDir = new URL('.', import.meta.url).pathname;
+// Convert the import.meta.url to a file path properly for Windows
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
-const dirPath = path.resolve(currentDir, '../dist/cdn');
+const dirPath = path.resolve(__dirname, '../dist/cdn');
 const fileToKeep = 'night-vision.min.js';
 
 export default function cleanDist() {
+  if (!fs.existsSync(dirPath)) return;
   fs.readdirSync(dirPath).forEach(file => {
     if (file !== fileToKeep) {
       fs.unlinkSync(path.join(dirPath, file));
